Updated the Deployment Recipe for Capistrano.
Updated Deployment Recipe

+++ files/deploy.rb
@@ -27,21 +27,27 @@
 #     cap deploy:apache:create
 #     cap deploy:apache:destroy
 #     cap deploy:apache:restart
+#     cap deploy:destroy_all:apache
 #
 #   For NginX Users
 #     cap deploy:nginx:create
 #     cap deploy:nginx:destroy
 #     cap deploy:nginx:restart
+#     cap deploy:destroy_all:nginx
 
 # Configuration
 
 # This configuration is *essential*
-set   :application,         "domain.com"
+
+set   :ip,                  "123.45.6785.910"
+set   :domain,              "domain.com"
+set   :subdomain,           false # unless domain is a subdomain, e.g. subdomain.domain.com
 set   :user,                "root"
 
 # This configuration is *conventional*
-set   :deploy_to,           "/var/rails/#{application}"
-set   :repository_path,     "/var/git/#{application}.git"
+set   :application,         ip
+set   :deploy_to,           "/var/rails/#{domain}"
+set   :repository_path,     "/var/git/#{domain}.git"
 set   :repository,          "ssh://#{user}@#{application}#{repository_path}"
 
 # The following configuration *optional*
@@ -53,16 +59,40 @@ role  :app,                 application
 role  :db,                  application
 default_run_options[:pty] = true 
 
-# Symlinks that should be created after each deployment
+# Symlinks
+# that should be created after each deployment
+#
 # The index[0] is the shared_path
 # The index[1] is the release_path
 symlink_configuration = [
-  ['config/database.yml',     'config/database.yml'   ],
-  ['db/production.sqlite3',   'db/production.sqlite3' ],
-  ['system',                  'public/system'         ],
-  ['assets',                  'public/assets'         ]
+  %w(config/database.yml    config/database.yml),
+  %w(db/production.sqlite3  db/production.sqlite3),
+  %w(system                 public/system)
 ]
 
+# Shared Folders
+# That should be created inside the shared_path
+directory_configuration = %w(db config system)
+
+
+
+
+
+
+#
+# Helper Methods
+#
+
+def create_tmp_file(contents)
+  system 'mkdir tmp'
+  file = File.new("tmp/#{domain}", "w")
+  file << contents
+  file.close
+end
+
+# 
+# Do Not Alter The Recipe Below!
+# 
 namespace :deploy do
   
   # Tasks that run after the (cap deploy:setup)
@@ -70,7 +100,9 @@ namespace :deploy do
   desc "Sets up the shared path"
   task :setup_shared_path do
     puts "\n\n=== Setting up the shared path! ===\n\n"
-    run "mkdir -p #{shared_path}/db #{shared_path}/assets #{shared_path}/system #{shared_path}/config"
+    directory_configuration.each do |directory|
+      run "mkdir -p #{shared_path}/#{directory}"
+    end
     system "cap deploy:db:update_yaml"
   end
   
@@ -86,7 +118,6 @@ namespace :deploy do
     system 'cap deploy:passenger:restart'
   end
  
- 
   # Deployment Tasks
   
   desc "Executes the initial procedures for deploying a Ruby on Rails Application."
@@ -97,6 +128,24 @@ namespace :deploy do
     system "cap deploy:db:create"
     system "cap deploy:db:migrate"
     system "cap deploy:passenger:restart"
+  end
+  
+  namespace :destroy_all do
+    
+    desc "Destroys Git Repository, Rails Environment and Apache2 Configuration."
+    task :apache do
+      system "cap deploy:repository:destroy"
+      run "rm -rf #{deploy_to}"
+      system "cap deploy:apache:destroy"
+    end
+
+    desc "Destroys Git Repository, Rails Environment and Nginx Configuration."
+    task :nginx do
+      system "cap deploy:repository:destroy"
+      run "rm -rf #{deploy_to}"
+      system "cap deploy:nginx:destroy"
+    end
+    
   end
   
   namespace :passenger do
@@ -131,14 +180,17 @@ namespace :deploy do
     end
   end
   
-
   # Manual Tasks
   
   namespace :db do
   
     desc "Syncs the database.yml file from the local machine to the remote machine"
-    task :update_yaml do
+    task :sync_yaml do
       puts "\n\n=== Syncing database yaml to the production server! ===\n\n"
+      unless File.exist?("config/database.yml")
+        puts "There is no config/database.yml.\n "
+        exit
+      end
       system "rsync -vr --exclude='.DS_Store' config/database.yml #{user}@#{application}:#{shared_path}/config/"
     end
     
@@ -174,16 +226,17 @@ namespace :deploy do
   end
   
   namespace :repository do
-    
+
     desc "Creates the remote Git repository."
     task :create do
       puts "\n\n=== Creating remote Git repository! ===\n\n"
       run "mkdir -p #{repository_path}"
       run "cd #{repository_path} && git --bare init"
+      system "git remote rm origin"
       system "git remote add origin #{repository[:repository]}"
       p "#{repository[:repository]} was added to your git repository as origin/master."
     end
-    
+
     desc "Creates the remote Git repository."
     task :destroy do
       puts "\n\n=== destroying remote Git repository! ===\n\n"
@@ -199,6 +252,13 @@ namespace :deploy do
       system "cap deploy:repository:create"
     end
     
+    desc "Reinitializes Origin/Master."
+    task :reinitialize do
+      system "git remote rm origin"
+      system "git remote add origin #{repository[:repository]}"
+      p "#{repository[:repository]} (origin/master) was added to your git repository."
+    end
+
   end
   
   namespace :environment do
@@ -218,7 +278,7 @@ namespace :deploy do
       run "rm -rf #{deploy_to}"
       system "cap deploy:setup"
     end
-  
+
   end
   
   namespace :crontab do
@@ -226,7 +286,7 @@ namespace :deploy do
     desc "Update the crontab file for the Whenever Gem."
     task :update, :roles => :db do
       puts "\n\n=== Updating the Crontab! ===\n\n"
-      run "cd #{release_path} && whenever --update-crontab #{application}"
+      run "cd #{release_path} && whenever --update-crontab #{domain}"
     end
     
   end
@@ -235,21 +295,22 @@ namespace :deploy do
     
     desc "Adds Apache2 configuration and enables it."
     task :create do
-      puts "\n\n=== Adding Apache2 Virtual Host for #{application}! ===\n\n"
+      puts "\n\n=== Adding Apache2 Virtual Host for #{domain}! ===\n\n"
       config = <<-CONFIG
       <VirtualHost *:80>
-        ServerName #{application}
-        ServerAlias www.#{application}
+        ServerName #{domain}
+        #{unless subdomain then "ServerAlias www.#{domain}" end}
         DocumentRoot #{File.join(deploy_to, 'current', 'public')}
       </VirtualHost>
       CONFIG
       
-      file = File.new("tmp/#{application}", "w")
+      system 'mkdir tmp'
+      file = File.new("tmp/#{domain}", "w")
       file << config
       file.close 
-      system "rsync -vr tmp/#{application} #{user}@#{application}:/etc/apache2/sites-available/#{application}"
-      File.delete("tmp/#{application}")
-      run "sudo a2ensite #{application}"
+      system "rsync -vr tmp/#{domain} #{user}@#{application}:/etc/apache2/sites-available/#{domain}"
+      File.delete("tmp/#{domain}")
+      run "sudo a2ensite #{domain}"
       run "sudo /etc/init.d/apache2 restart"
     end
     
@@ -260,9 +321,9 @@ namespace :deploy do
     
     desc "Removes Apache2 configuration and disables it."
     task :destroy do
-      puts "\n\n=== Removing Apache2 Virtual Host for #{application}! ===\n\n"
-      begin run("a2dissite #{application}"); rescue; end
-      begin run("sudo rm /etc/apache2/sites-available/#{application}"); rescue; end
+      puts "\n\n=== Removing Apache2 Virtual Host for #{domain}! ===\n\n"
+      begin run("a2dissite #{domain}"); rescue; end
+      begin run("sudo rm /etc/apache2/sites-available/#{domain}"); rescue; end
       run("sudo /etc/init.d/apache2 restart")
     end
 
@@ -272,21 +333,20 @@ namespace :deploy do
     
     desc "Adds NginX configuration and enables it."
     task :create do
-      puts "\n\n=== Adding NginX Virtual Host for #{application}! ===\n\n"
+      puts "\n\n=== Adding NginX Virtual Host for #{domain}! ===\n\n"
       config = <<-CONFIG
       server {
         listen 80;
+        server_name #{unless subdomain then "www.#{domain} #{domain}" else domain end};
         root #{File.join(deploy_to, 'current', 'public')};
         passenger_enabled on;
       }
       CONFIG
-            
-      file = File.new("tmp/#{application}", "w")
-      file << config
-      file.close
+      
+      create_tmp_file(config)
       run "mkdir -p /opt/nginx/conf/sites-enabled"
-      system "rsync -vr tmp/#{application} #{user}@#{application}:/opt/nginx/conf/sites-enabled/#{application}"
-      File.delete("tmp/#{application}")
+      system "rsync -vr tmp/#{domain} #{user}@#{application}:/opt/nginx/conf/sites-enabled/#{domain}"
+      File.delete("tmp/#{domain}")
       system 'cap deploy:nginx:restart'
     end
     
@@ -298,9 +358,9 @@ namespace :deploy do
     
     desc "Removes NginX configuration and disables it."
     task :destroy do
-      puts "\n\n=== Removing NginX Virtual Host for #{application}! ===\n\n"
+      puts "\n\n=== Removing NginX Virtual Host for #{domain}! ===\n\n"
       begin
-        run("rm /opt/nginx/conf/sites-enabled/#{application}")
+        run("rm /opt/nginx/conf/sites-enabled/#{domain}")
       ensure
         system 'cap deploy:nginx:restart'
       end
